OpenSSL support for TLS offload.
This patch adds support to our in-tree OpenSSL to support
TLS offload in the transmit direction only.  All apps linking
to the library using normal SSL apis will be accelerated in
the transmit direction, as long as kernel TLS support is
present, and they are using a cipher supported by the kerne.

This patch also adds a new SSL_sendfile() call.

The of placement of the transition to kernel offload in the openssl
libs was inspired by Boris Pismenny <borisp@mellanox.com>'s work
on Linux kTLS.

rrs wrote the initial version of this support, including the
SSL_sendfile() api.

Sponsored by: Netflix