Add an SB_TLS_IFNET flag for sb_tls_flags.
When this flag is set, encryption is performed in the ifnet layer rather
than above the protocol layer.  Don't call sbtls_encrypt() when this flag
is set, don't mark mbufs as M_NOTREADY, and don't pass PRUS_NOTREADY to
pru_send() when this flag is set on a socket buffer.

This required changing sbtls_frame() to accept the socket instead of the
tls structure so that it can see the state of SB_TLS_IFNET flag.