Add a facility for transmitting "raw" work requests on regular NIC queues.
Use PH_loc.eight[2] as a new flag to indicate that an mbuf is a
work-request (WR) mbuf rather than an Ethernet packet.  One of these
mbufs is allocated via alloc_wr_mbuf() which will allocate a single
contiguous range to hold the mbuf data.  The consumer can use mtod()
to obtain the start of the work request and write the required work
request in the buffer.  The mbuf can then be enqueued directly to the
txq via mp_ring_enqueue().