Extend the T6 cipher structure with two callback functions.
The first is used in parse_pkt() to calculate 'nsegs' and 'len16' for
a TLS packet.  The second is used to write out the actual work request
into the transmit ring.  An initial version of the first callback is
implemented to calculate the full work request size.  The second
callback is just a stub currently.

Right now the first callback always reserves room in 'len16' for up to
5 CPL_SET_TCB_FIELD requests in the txq.  Note that since the second
callback returns the actual number of descriptors used, the second
callback can choose to omit some of those work requests if they are
not needed.