Don't modify the TLS header in the ext_pgs structure.
The TLS header bytes need to remain immutable so that retransmits see the
proper (shortened) length for each retransmit.  To handle this, use a gross
hack and abuse the empty space at the end of the header mbuf (after the
TCP header) to store the TLS header.  Be careful to always store the exact
same info (and to write it out in such a way that a retransmit won't change
the data).