Initial version of supporting "short" TLS records.
Look at the TLS record mbuf to determine how much of the TLS record
the stack is actually wanting to transmit.  If the stack has asked to
not transmit the end of the TLS record, use AES-CTR to encrypt the
partial payload.  Note that we always start the TLS record from the
beginning even if the mbuf only requests a middle range.

The TLS header is passed through as "header" data in the SCMD0 HdrLength
field rather than as AAD.  A suitable IV for AES-CTR is computed from
the salt, TLS sequence number, and an initial counter of 2 and placed
after the TLS header but before the TLS record plaintext.  AADIVDROP is
set in SCMD0 so that the IV is not included in the resulting packet
data.  I was able to reuse the AES-GCM key context since AES-CTR only
needs the AES key and ignores the GHASH portion of the key context.