Improve handling of work request credits.
Previously I cheated a bit and stored a single count of the combined
work requests (0 to 3 CPL_SET_TCB_FIELDs followed by 1 or more TLS
WRs) in the first software descriptor (txsd).  However, the firmware
could acknowledge a subset of these work requests in the status page
which then triggered an assertion failure (it looked like the firmware
had returned partial credits).  To fix, properly account for the
credits used by each individual WR in the matching software descriptor
(txsd) but only store the mbuf pointer in the final WR's txsd to
ensure the mbuf chain is only released once all of the WRs are
serviced.