Rework the hack to avoid losing TLS records in tcp_output.
With nginx, the first TLS record (FINISH) can be appended to the
socket buffer while other data is still waiting to be sent.  As a
result, m_copydata() was used in tcp_output() because the existing
check for a ifnet TLS record only looked at the head of the the socket
buffer.  Instead, ifnet TLS records always need to be sent down as-is
so that the driver can get to the TLS record.

The approach used is to force out the earlier handshake data as a
separate packet before the first TLS record.  This is perhaps not
ideal in terms of packets on the wire, but it only happens once per
connection at the end of the handshake.  Unfortunately, the check that
is used is a bit expensive and applied to all non-TLS mbufs.

While here, include #if 0'd code to try to expand packets to cover full
TLS records, though it doesn't work in my testing (hence the #if 0).