Correct condition for a "short" TLS record in the parse_pkt callback.
Previously this would not correctly account for the IV in the WR length
in the len16 saved in the mbuf.  If the actual WR needed a second descriptor
just due to the IV and there was exactly room in the ring for the "wrong"
len16 number of descriptors, this could cause us to overflow the TX ring
which was caught by an assertion failure.