Discard segments from the start of a TLS record if requested.
The input to the crypto engine must always begin at the start of a TLS
record.  However, individual MSS segments generated by CPL_TX_DATA
will be dropped instead of placed on the wire if their entire sequence
range is before TP's notion of SND_UNA.  Make use of this to set TX_MAX
and SND_UNA to the first byte we want to send.  This fixes requests that
ask to only submit the end of a TLS record to only submit the final few
MSS segments instead of the entire TLS record greatly reducing the
duplicate data.